backend_python/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.2.0
Pillow==11.0.0
